+++ interface/ceed-preconditioning.c
@@ -71,8 +71,8 @@ int CeedOperatorCreateFallback(CeedOperator op) {
   ierr = CeedCalloc(1, &op_ref); CeedChk(ierr);
   memcpy(op_ref, op, sizeof(*op_ref));
   op_ref->data = NULL;
-  op_ref->interface_setup = false;
-  op_ref->backend_setup = false;
+  op_ref->is_interface_setup = false;
+  op_ref->is_backend_setup = false;
   op_ref->ceed = ceed_ref;
   ierr = ceed_ref->OperatorCreate(op_ref); CeedChk(ierr);
   op->op_fallback = op_ref;
@@ -451,7 +451,7 @@ static int CeedSingleOperatorAssembleSymbolic(CeedOperator op, CeedInt offset,
     CeedInt *rows, CeedInt *cols) {
   int ierr;
   Ceed ceed = op->ceed;
-  if (op->composite)
+  if (op->is_composite)
     // LCOV_EXCL_START
     return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
                      "Composite operator not supported");
@@ -539,7 +539,7 @@ static int CeedSingleOperatorAssemble(CeedOperator op, CeedInt offset,
                                       CeedVector values) {
   int ierr;
   Ceed ceed = op->ceed;
-  if (op->composite)
+  if (op->is_composite)
     // LCOV_EXCL_START
     return CeedError(ceed, CEED_ERROR_UNSUPPORTED,
                      "Composite operator not supported");
@@ -799,7 +799,7 @@ static int CeedSingleOperatorAssemblyCountEntries(CeedOperator op,
   CeedElemRestriction rstr;
   CeedInt num_elem, elem_size, num_comp;
 
-  if (op->composite)
+  if (op->is_composite)
     // LCOV_EXCL_START
     return CeedError(op->ceed, CEED_ERROR_UNSUPPORTED,
                      "Composite operator not supported");
@@ -1036,6 +1036,9 @@ CeedPragmaOptimizeOn
     consists of (1 + dim) x (dim + 1) matrices at each quadrature point acting
     on the input [u, du_0, du_1] and producing the output [dv_0, dv_1, v].
 
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
+
   @param op              CeedOperator to assemble CeedQFunction
   @param[out] assembled  CeedVector to store assembled CeedQFunction at
                            quadrature points
@@ -1077,6 +1080,9 @@ int CeedOperatorLinearAssembleQFunction(CeedOperator op, CeedVector *assembled,
 
   Note: Currently only non-composite CeedOperators with a single field and
           composite CeedOperators with single field sub-operators are supported.
+
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
 
   @param op              CeedOperator to assemble CeedQFunction
   @param[out] assembled  CeedVector to store assembled CeedOperator diagonal
@@ -1132,6 +1138,9 @@ int CeedOperatorLinearAssembleDiagonal(CeedOperator op, CeedVector assembled,
 
   Note: Currently only non-composite CeedOperators with a single field and
           composite CeedOperators with single field sub-operators are supported.
+
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
 
   @param op              CeedOperator to assemble CeedQFunction
   @param[out] assembled  CeedVector to store assembled CeedOperator diagonal
@@ -1191,6 +1200,9 @@ int CeedOperatorLinearAssembleAddDiagonal(CeedOperator op, CeedVector assembled,
 
   Note: Currently only non-composite CeedOperators with a single field and
           composite CeedOperators with single field sub-operators are supported.
+
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
 
   @param op              CeedOperator to assemble CeedQFunction
   @param[out] assembled  CeedVector to store assembled CeedOperator point block
@@ -1254,6 +1266,9 @@ int CeedOperatorLinearAssemblePointBlockDiagonal(CeedOperator op,
 
   Note: Currently only non-composite CeedOperators with a single field and
           composite CeedOperators with single field sub-operators are supported.
+
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
 
   @param op              CeedOperator to assemble CeedQFunction
   @param[out] assembled  CeedVector to store assembled CeedOperator point block
@@ -1324,6 +1339,9 @@ int CeedOperatorLinearAssembleAddPointBlockDiagonal(CeedOperator op,
    ordering.
 
    This will generally be slow unless your operator is low-order.
+
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
 
    @param[in]  op           CeedOperator to assemble
    @param[out] num_entries  Number of entries in coordinate nonzero pattern
@@ -1417,6 +1435,9 @@ int CeedOperatorLinearAssembleSymbolic(CeedOperator op, CeedInt *num_entries,
 
    This will generally be slow unless your operator is low-order.
 
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
+
    @param[in]  op      CeedOperator to assemble
    @param[out] values  Values to assemble into matrix
 
@@ -1477,6 +1498,9 @@ int CeedOperatorLinearAssemble(CeedOperator op, CeedVector values) {
            for a CeedOperator, creating the prolongation basis from the
            fine and coarse grid interpolation
 
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
+
   @param[in] op_fine       Fine grid operator
   @param[in] p_mult_fine   L-vector multiplicity in parallel gather/scatter
   @param[in] rstr_coarse   Coarse grid restriction
@@ -1495,6 +1519,7 @@ int CeedOperatorMultigridLevelCreate(CeedOperator op_fine,
                                      CeedOperator *op_coarse, CeedOperator *op_prolong,
                                      CeedOperator *op_restrict) {
   int ierr;
+  ierr = CeedOperatorCheckReady(op_fine); CeedChk(ierr);
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op_fine, &ceed); CeedChk(ierr);
 
@@ -1584,6 +1609,9 @@ int CeedOperatorMultigridLevelCreate(CeedOperator op_fine,
   @brief Create a multigrid coarse operator and level transfer operators
            for a CeedOperator with a tensor basis for the active basis
 
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
+
   @param[in] op_fine        Fine grid operator
   @param[in] p_mult_fine    L-vector multiplicity in parallel gather/scatter
   @param[in] rstr_coarse    Coarse grid restriction
@@ -1602,6 +1630,7 @@ int CeedOperatorMultigridLevelCreateTensorH1(CeedOperator op_fine,
     const CeedScalar *interp_c_to_f, CeedOperator *op_coarse,
     CeedOperator *op_prolong, CeedOperator *op_restrict) {
   int ierr;
+  ierr = CeedOperatorCheckReady(op_fine); CeedChk(ierr);
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op_fine, &ceed); CeedChk(ierr);
 
@@ -1651,6 +1680,9 @@ int CeedOperatorMultigridLevelCreateTensorH1(CeedOperator op_fine,
   @brief Create a multigrid coarse operator and level transfer operators
            for a CeedOperator with a non-tensor basis for the active vector
 
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
+
   @param[in] op_fine        Fine grid operator
   @param[in] p_mult_fine    L-vector multiplicity in parallel gather/scatter
   @param[in] rstr_coarse    Coarse grid restriction
@@ -1673,6 +1705,7 @@ int CeedOperatorMultigridLevelCreateH1(CeedOperator op_fine,
                                        CeedOperator *op_prolong,
                                        CeedOperator *op_restrict) {
   int ierr;
+  ierr = CeedOperatorCheckReady(op_fine); CeedChk(ierr);
   Ceed ceed;
   ierr = CeedOperatorGetCeed(op_fine, &ceed); CeedChk(ierr);
 
@@ -1729,6 +1762,9 @@ int CeedOperatorMultigridLevelCreateH1(CeedOperator op_fine,
     diagonalization and obtain an approximate inverse of the form
       V^T S^hat V. The CeedOperator must be linear and non-composite. The
     associated CeedQFunction must therefore also be linear.
+
+  Note: Calling this function asserts that setup is complete
+          and sets the CeedOperator as immutable.
 
   @param op            CeedOperator to create element inverses
   @param[out] fdm_inv  CeedOperator to apply the action of a FDM based inverse